--- docs/ChangeLog.rst
@@ -1,64 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.2.2 (2024-06-27)
--------------------
-
-**功能更新**
-1. 新增`bdp.py`脚本，目前已包含`dl.py` `bd.py`在内的所有功能，在此基础上修改了输入方式。如果你想要直接通过链接下载，推荐使用`bd.py`。`bdp.py`仅用于需要下载多个链接中的所有表情包。
-#. 在使用`bdp.py`时，将在当前目录下生成`output`文件夹，所有的表情包都将下载在该文件夹中。
-#. 使用方法见`README文档`。在使用`bdp.py`时，请一定在与脚本相同的目录下创建`bdp.txt`文件。
-
-**Feature Updates**
-1. Added `bdp.py` script, which now includes all the functionalities of `dl.py` and `bd.py`. The input method has been modified based on this. If you want to download directly from a link, it's recommended to use `bd.py`. `bdp.py` is only used for downloading all the sticker packs from multiple links.
-#. When using `bdp.py`, an `output` folder will be created in the current directory, and all the sticker packs will be downloaded into this folder.
-#. For usage, please refer to the `README document`. When using `bdp.py`, make sure to create a `bdp.txt` file in the same directory as the script.
-
-.. _README文档: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
-
-.. _README document: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
-
-v0.2.1 (2024-06-09)
--------------------
-
-**功能更新**
-1. 修改了动态表情包和静态表情包的下载顺序，减少下载请求。
-#. 整合了批量下载的功能，详情请见 `README.md`。
-#. 使用大小驼峰法减少了代码冗余。
-#. 修改了正则表达式使其更加兼容LINE的表情包规则。
-#. 学习了非常好markdown使我的README旋转。
-**Feature Updates**
-1. Modified the download order of animated and static stickers to reduce download requests.
-#. Integrated batch download functionality. For details, please refer to `README.md`.
-#. Reduced code redundancy using upper camel case notation.
-#. Modified the regular expression to make it more compatible with LINE's sticker rules.
-
-.. _README.md: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
-
-v0.1.2(2024-05-16)
--------------------
-
-This version adds a feature to delete key thumbnails from zip files, saving me the trouble of repeatedly deleting key files and greatly improving my work efficiency!
-In addition, since the script written by ChatGPT-4o still has bugs, I have manually modified some of the code, so the current code is no longer completely AI-generated, but is now **partially AI-generated!**
-However, this version refers to `shutil` `zipfile` `requests` and `BeautifulSoup4` dependencies, and you can install the dependencies by the following command.
-```sh
-python -m pip install requests beautifulsoup4 zipfile shutil
-```
-
-v0.1.1(2024-05-16)
--------------------
-
-This version adds the function of crawling the name of the expression pack from the web page and renaming it, so now you won't encounter the problem of batch downloading because the names are all the same!
-However, this version refers to `requests` and `BeautifulSoup` dependencies, and you can install the dependencies by the following command.
-```sh
-python -m pip install BeautifulSoup requests
-```
-**Generated by AI**
-
-v0.1.0(2024-05-16)
--------------------
-
-This version distinguishes emoji from sticker, and then uses regular expressions to match the ID in the url and download it. The script is very elementary, but I hope you can enjoy it.
-Generated through AI
