+++ docs/README_EN.md
@@ -2,7 +2,7 @@
 
 # LINE Stickers and Emoji Downloader
 
-This project can be used to download LINE stickers and emoji, including animated and static images, and batch download operations. All the python script content is generated by AI (chatGPT-4o, Gemini1.5PRO).  
+This project can be used to download LINE stickers and emoji, including animated and static images, and batch download operations. All the python script content is generated by AI (chatGPT-4o, Gemini1.5PRO).
 This project is prohibited for illegal use and sale. **If you like these stickers, please support the official LINE stickers**
 
 ## Directory
@@ -22,44 +22,52 @@ This project is prohibited for illegal use and sale. **If you like these sticker
 
 ## Script Dependencies
 
-> First, make sure that your computer has **python3.6** or above installed and python is included in your computer's **PATH**! Ensure your pip is the latest version, you can update your pip by running `python -m pip install --upgrade pip`.
+> First, make sure that your computer has **python3.11** or above installed and python is included in your computer's **PATH**! Ensure your pip is the latest version, you can update your pip by running `python -m pip install --upgrade pip`.
 
-You can run the following command to install the dependent libraries.
+Run the following command to install the dependent libraries:
+
 ```bash
-pip install BeautifulSoup4 zipfile django
+pip install -r requirements.txt
 ```
 
 ## How to use
 
-| File name | Function                                                                | How to use                                                       |
-|-----------|-------------------------------------------------------------------------|------------------------------------------------------------------|
-| `dl.py`   | This file only has the function of downloading a single sticker pack    | Please copy the sticker pack URL `url` you want to download, and you can run `python dl.py <url>` in the same directory to download.                                  |
-| `bd.py`   | Download a single sticker or batch download all stickers from a creator | Please copy the sticker pack URL or creator URL you want to download, and enter `python bd.py` in the same directory and run it, paste the URL after the prompt. |
-| `bdp.py`  | Similar to `bd.py`                                                      | Paste the URLs of the texture packs you want to download and the author's URLs into `bdp.txt` in the same directory as the script, separated by carriage returns. See [here](./docs/bdp.txt) for an example. Then run `python bdp.py` and the script will automatically download the textures in bulk.                                     |
+| File Name    | Function                                                             | Usage                                                                                                                                                                                                                                                                                                                                                                     |
+|--------------|----------------------------------------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| `main.py`    | Three modes:<br>1. Download a single sticker pack<br>2. Download all stickers by an author<br>3. Batch download from a txt file | Run `python main.py` in terminal and select a mode as prompted:<br> • **Mode 1**: Enter `1`, then choose whether to download to a specific folder (type `y` and enter the path, or `n` to use the current directory). Paste the URL of a single sticker/emoji product. The script will download and automatically remove any `key` files from the zip.<br> • **Mode 2**: Enter `2`, similarly choose the download folder first, then paste the author’s homepage URL (e.g., `https://store.line.me/author/xxx`). The script will iterate through all pages and download all sticker packs by that author.<br> • **Mode 3**: Enter `3`, choose the download folder, then enter the path to a txt file (press Enter to use `bdp.txt` in the current directory). Each line in the txt file should contain a URL (product or author link). The script will process them in batch. |
+
+- Supported types: LINE regular stickers (`/stickershop/product/`) and emojis (`/emojishop/product/`).
+- Downloaded zip files are named after the sticker pack, and any internal `key` files are automatically removed.
+- If you choose to download to a specific folder, it will be created automatically if it does not exist.
 
 ## Errors and solutions
 
 ### 1. Unable to get the sticker pack name
+
 If the downloaded file is named `unknown_emoji_name`, please check your network environment and whether your network IP is outside the region where LINE provides services.
 
 ### 2. Unable to download
+
 If the program runs an error and leaves an empty folder in the directory, please raise an issue and try to reproduce the content. The problem has occurred before, but the problem was solved and forgotten how to reproduce the problem.
 
 ### 3. Temp files are not deleted after download
+
 - If you use termux on your mobile phone and try to download in the storage directory, then you have stacked the buffs. The termux on some mobile phones intercepts the deletion of files in the storage directory and its subdirectories.
 - If you are using another device or the error does not occur in the storage directory, please check if you have permissions in the directory of this folder.
 
 ### 4. Incomplete Downloads
+
 - When the number of emoji packs uploaded by the creator exceeds 36, only the emoji packs on a single page can be sniffed. Please use `bdp.py` to download, and paste all the URLs of the creator's emoji pack pages into `bdp.txt`.
 - This issue has been fixed in version `v0.2.4` and later.
 
 ### 5. Others
+
 If you encounter other problems besides the above, please raise an issue and reproduce the problem record submitted. Thank you for your support of this project.
 
 ## Acknowledgement
 
-Thanks to the following friends and contributors: (in no particular order)  
-[@kaixinol](https://github.com/kaixinol) | [@ZGQ Inc.](https://github.com/ZGQ-inc) | [@CPuddingOwO](https://github.com/CPuddingOwO) 
+Thanks to the following friends and contributors: (in no particular order)
+[@kaixinol](https://github.com/kaixinol) | [@ZGQ Inc.](https://github.com/ZGQ-inc) | [@CPuddingOwO](https://github.com/CPuddingOwO)
 
 ## License
