--- realease/v0.1.0/ChangeLog.rst
@@ -1,10 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.1.0(2024-05-16)
--------------------
-
-This version distinguishes emoji from sticker, and then uses regular expressions to match the ID in the url and download it. The script is very elementary, but I hope you can enjoy it.
-Generated through AI
