--- realease/v0.1.1/ChangeLog.rst
@@ -1,14 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.1.1(2024-05-16)
--------------------
-
-This version adds the function of crawling the name of the expression pack from the web page and renaming it, so now you won't encounter the problem of batch downloading because the names are all the same!
-However, this version refers to `requests` and `BeautifulSoup` dependencies, and you can install the dependencies by the following command.
-```sh
-python -m pip install BeautifulSoup requests
-```
-**Generated by AI**
