--- realease/v0.1.2/ChangeLog.rst
@@ -1,14 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.1.2(2024-05-16)
--------------------
-
-This version adds a feature to delete key thumbnails from zip files, saving me the trouble of repeatedly deleting key files and greatly improving my work efficiency!
-In addition, since the script written by ChatGPT-4o still has bugs, I have manually modified some of the code, so the current code is no longer completely AI-generated, but is now **partially AI-generated!**
-However, this version refers to `shutil` `zipfile` `requests` and `BeautifulSoup4` dependencies, and you can install the dependencies by the following command.
-```sh
-python -m pip install requests beautifulsoup4 zipfile shutil
-```
