--- realease/v0.2.1/ChangeLog.rst
@@ -1,21 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.2.1 (2024-06-09)
--------------------
-
-**功能更新**
-1. 修改了动态表情包和静态表情包的下载顺序，减少下载请求。
-#. 整合了批量下载的功能，详情请见 `README.md`。
-#. 使用大小驼峰法减少了代码冗余。
-#. 修改了正则表达式使其更加兼容LINE的表情包规则。
-#. 学习了非常好markdown使我的README旋转。
-**Feature Updates**
-1. Modified the download order of animated and static stickers to reduce download requests.
-#. Integrated batch download functionality. For details, please refer to `README.md`.
-#. Reduced code redundancy using upper camel case notation.
-#. Modified the regular expression to make it more compatible with LINE's sticker rules.
-
-.. _README.md: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
