--- realease/v0.2.2/ChangeLog.rst
@@ -1,21 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.2.2 (2024-06-27)
--------------------
-
-**功能更新**
-1. 新增`bdp.py`脚本，目前已包含`dl.py` `bd.py`在内的所有功能，在此基础上修改了输入方式。如果你想要直接通过链接下载，推荐使用`bd.py`。`bdp.py`仅用于需要下载多个链接中的所有表情包。
-#. 在使用`bdp.py`时，将在当前目录下生成`output`文件夹，所有的表情包都将下载在该文件夹中。
-#. 使用方法见`README文档`。在使用`bdp.py`时，请一定在与脚本相同的目录下创建`bdp.txt`文件。
-
-**Feature Updates**
-1. Added `bdp.py` script, which now includes all the functionalities of `dl.py` and `bd.py`. The input method has been modified based on this. If you want to download directly from a link, it's recommended to use `bd.py`. `bdp.py` is only used for downloading all the sticker packs from multiple links.
-#. When using `bdp.py`, an `output` folder will be created in the current directory, and all the sticker packs will be downloaded into this folder.
-#. For usage, please refer to the `README document`. When using `bdp.py`, make sure to create a `bdp.txt` file in the same directory as the script.
-
-.. _README文档: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
-
-.. _README document: https://github.com/djmh1793225009/LINE_sticker_emoji_downloader/blob/main/README.md
