--- realease/v0.2.3/ChangeLog.rst
@@ -1,21 +1,0 @@
-.. _changelog:
-
-ChangeLog
-==========
-
-v0.2.3 (2024-07-03)
--------------------
-
-**功能更新**
-1. 在`bd.py`脚本中添加**是否新建文件夹**功能。
-
-**代码优化**
-1. 优化代码块和函数名，减少了代码冗余。
-2. 改变了库调用，减少依赖项。不再需要`shutil`库依赖。
-
-**Function Updates**
-1. Add **whether to create a new folder** functionality in the `bd.py` script.
-
-**Code Optimization**
-1. Optimized code blocks and function names, reducing code redundancy.
-2. Changed library calls, reducing dependencies. No longer requires the `shutil` library.
