+++ requirements.txt
@@ -1,0 +1,3 @@
+django
+BeautifulSoup4
+requests
